+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-Bugfixes
+# Changelog
+* Fixed a bug where flight power would go down in spectator mode